+++ src/styles/app.scss
@@ -6,4 +6,4 @@
 @import './header';
 @import './footer';
 @import './contact';
-@import './stylesheet';
+@import './global-stylesheet';
